requirements_dev.txt
pre-commit
lib-elro-connects==0.5.1
homeassistant
pytest-homeassistant-custom-component
